Refatoração de camadas da API, dto implementado e mais exception handler

+++ src/main/java/com/francinjr/xpenses/domain/service/GanhoService.java
@@ -9,6 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.francinjr.xpenses.domain.exception.GanhoNaoEncontradoException;
 import com.francinjr.xpenses.domain.model.Ganho;
 import com.francinjr.xpenses.domain.repository.GanhoRepository;
+import com.francinjr.xpenses.dto.GanhoDTO;
 
 @Service
 public class GanhoService {
@@ -16,27 +17,63 @@ public class GanhoService {
 	@Autowired
 	private GanhoRepository ganhoRepository;
 
-	public List<Ganho> buscarTodos() {
-		return ganhoRepository.findAll();
+	public List<GanhoDTO> buscarTodos() {
+		List<Ganho> ganhos = ganhoRepository.findAll();
+
+		return ganhos.stream().map(ganho -> new GanhoDTO(ganho.getNome(), ganho.getValor(), 
+				ganho.getDescricao(), ganho.getId())).toList();
+		
 	}
 	
-	public Ganho buscarOuFalhar(Long ganhoId) {
-		return ganhoRepository.findById(ganhoId)
+	public GanhoDTO buscarOuFalhar(Long ganhoId) {
+		Ganho ganho = ganhoRepository.findById(ganhoId)
 				.orElseThrow(() -> new GanhoNaoEncontradoException(ganhoId));
+		
+		GanhoDTO ganhoDTO = new GanhoDTO(ganho.getNome(), ganho.getValor(), 
+				ganho.getDescricao(), ganho.getId());
+		
+		return ganhoDTO;
 	}
 	
 	@Transactional
-	public Ganho salvar(Ganho ganho) {
-		return ganhoRepository.save(ganho);
+	public GanhoDTO criar(GanhoDTO data) throws Exception {
+		Ganho ganho = new Ganho(data);
+		
+		Ganho ganhoCriado = ganhoRepository.save(ganho);
+		
+		GanhoDTO ganhoSalvo = new GanhoDTO(ganhoCriado.getNome(), ganhoCriado.getValor(), 
+				ganhoCriado.getDescricao(), ganhoCriado.getId());
+		
+		return ganhoSalvo;
+	}
+	
+	@Transactional
+	public GanhoDTO atualizar(GanhoDTO data) throws Exception {
+		//Ganho ganho = new Ganho(data);
+		GanhoDTO ganhoBuscado = buscarOuFalhar(data.id());
+		
+		if(ganhoBuscado != null) {
+			Ganho ganhoEncontrado = new Ganho(data);
+			ganhoEncontrado = ganhoRepository.save(ganhoEncontrado);
+			
+			GanhoDTO ganhoSalvo = new GanhoDTO(ganhoEncontrado.getNome(), 
+					ganhoEncontrado.getValor(), ganhoEncontrado.getDescricao(), 
+					ganhoEncontrado.getId());
+			
+			return ganhoSalvo;
+		} else {
+			throw new GanhoNaoEncontradoException(data.id());
+		}
 	}
 
 	@Transactional
 	public void excluir(Long ganhoId) {
-		try {
+		GanhoDTO ganhoBuscado = buscarOuFalhar(ganhoId);
+		
+		if(ganhoBuscado != null) {
 			ganhoRepository.deleteById(ganhoId);
-			//ganhoRepository.flush();
-		} catch (Exception e) {
-			e.printStackTrace();
+		} else {
+			throw new GanhoNaoEncontradoException(ganhoId);
 		}
 	}
 }
